requirements.txt
fastapi
uvicorn[standard]
httpx
ollama
pyyaml
python-dotenv
websockets